requirements.txt
# Core ML/AI Libraries
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
transformers==4.30.2

# Web Framework
Flask==2.3.2
Flask-CORS==4.0.0
Werkzeug==2.3.6

# System Monitoring
psutil==5.9.5
GPUtil==1.4.0

# Data Processing & Storage
joblib==1.3.1
pickle-mixin==1.0.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# ML Model Tracking & Versioning
mlflow==2.4.1

# Explainability
shap==0.42.1
lime



# Time-series Analysis
Prophet==1.1.5
statsmodels==0.14.0

# Database (for time-series metrics)
influxdb-client==1.18.0
# OR
psycopg2-binary==2.9.6  # For PostgreSQL/TimescaleDB

# Hyperparameter Tuning
optuna==3.3.0

# Data Validation
pydantic==1.10.12

# Logging & Monitoring
python-dotenv==1.0.0
colorlog==6.7.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# API Documentation
flask-restx==0.5.1

# Async Task Queue
celery==5.3.1
redis==4.6.0

# Config Management
PyYAML==6.0

# Security
cryptography==41.0.0

# Development
black==23.7.0
pylint==2.17.5
flake8==6.0.0
